requirements.txt
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
chardet==4.0.0
et-xmlfile==1.0.1
idna==2.10
numpy==1.19.5
openpyxl==3.0.7
pandas==1.1.5
python-dateutil==2.8.1
python-dotenv==0.16.0
pytz==2021.1
requests==2.25.1
six==1.15.0
soupsieve==2.2.1
urllib3==1.26.4
xlrd==2.0.1
